Valid user model with password added
User model with controller and views added.  Validation is present for
user information.  Validates that the password is at least 6 characters
long, there is a name that consists of a string that is a maximum of 50
characters long, there must also be a valid email address which is not
case sensitive and checks with valid email regex, the add a new user
does not technically work because no password field is available to
enter a password in